Move all type declarations iinto a declaration file, remove explicit imports
